requirements.txt
appdirs
telethon-aio
tqdm
PySocks